messenger/requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
astroid==2.12.12
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.2
django-filter==22.1
django-login-required-middleware==0.9.0
djangorestframework==3.14.0
executing==1.1.1
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
kombu==5.2.4
lazy-object-proxy==1.8.0
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.5.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
astroid==2.12.12
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.2
django-filter==22.1
django-login-required-middleware==0.9.0
djangorestframework==3.14.0
executing==1.1.1
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
ipython==8.5.0
isort==5.10.1
jedi==0.18.1
kombu==5.2.4
lazy-object-proxy==1.8.0
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.5.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.5.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
